\name{HPC4Stats-package}
\alias{HPC4Stats-package}
\alias{HPC4Stats}
\docType{package}
\title{
\packageTitle{HPC4Stats}
}
\description{
\packageDescription{HPC4Stats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HPC4Stats}
\packageIndices{HPC4Stats}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{HPC4Stats}

Maintainer: \packageMaintainer{HPC4Stats}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
