\name{HPC4Stats-package}
\alias{HPC4Stats-package}
\alias{HPC4Stats}
\docType{package}
\title{
\packageTitle{HPC4Stats}
}
\description{
\packageDescription{HPC4Stats}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{HPC4Stats}
\packageIndices{HPC4Stats}
}
\author{
\packageAuthor{HPC4Stats}
Maintainer: \packageMaintainer{HPC4Stats}
}


